refactor: utils foldered
