+++ webapps/svelte/src/lib/app/AuthState.ts
@@ -1,5 +1,5 @@
 import AppSettings from "../../AppSettings";
-import { AuthTokens } from "../types";
+import { AuthTokens } from "../utils/types";
 import LoginAppComponent from "./LoginAppComponent";
 
 /**
