+++ webapps/svelte/src/lib/utils/helpers.ts
